+++ httpie_websockets.py
@@ -1,20 +1,44 @@
-import asyncio
 import io
+import json
+import logging
+import os
 import platform
+import queue
+import ssl
+import struct
 import sys
+import tempfile
 import threading
-from concurrent.futures import ThreadPoolExecutor
-from typing import AnyStr, AsyncIterable, Iterable, Optional
+import time
+from pathlib import Path
+from typing import Any, Mapping, Optional, Union
+from urllib.parse import ParseResult, urlparse
 
-import websockets
+import websocket
+from httpie.client import DEFAULT_UA
 from httpie.plugins import TransportPlugin
+from httpie.ssl_ import HTTPieCertificate
 from requests.adapters import BaseAdapter
 from requests.models import PreparedRequest, Response
 from requests.structures import CaseInsensitiveDict
+from websocket import ABNF, STATUS_ABNORMAL_CLOSED
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"
 __author__ = "belingud"
 __license__ = "MIT"
+
+logger = logging.getLogger(__name__)
+logging.basicConfig(
+    format="%(asctime)s %(levelname)s %(filename)s:%(lineno)s: %(message)s",
+    datefmt="%Y-%m-%d %H:%M:%S",
+)
+logger.setLevel(os.getenv("HTTPIE_WS_LOG_LEVEL", "WARNING").upper())
+
+try:
+    import python_socks  # noqa: F401
+except ImportError:
+    logger.debug("pyton-socks not installed, websocket proxy will not work")
+
 if platform.system().lower() == "windows":
     import msvcrt
 
@@ -39,186 +63,274 @@ else:
         return None
 
 
+def normalize_url(url, default_scheme="http") -> ParseResult:
+    """
+    Normalize a URL by adding `http` as default if it's missing.
+    Supported schemes: http, socks4, socks4a, socks5, socks5h
+    Examples:
+        'example.com' -> 'http://example.com'
+        '//example.com' -> 'http://example.com'
+        '://example.com' -> 'http://example.com'
+        'http://example.com' -> 'http://example.com'
+        'socks5://example.com' -> 'socks5://example.com'
+        'https://example.com' -> 'https://example.com'
+
+    Args:
+        url (str): The URL to normalize.
+        default_scheme (str, optional): The default scheme to use if the URL doesn't have one. Defaults to "http".
+
+    Returns:
+        ParseResult: The parsed URL.
+    """
+    if "." not in url:
+        try:
+            url.index("localhost")
+        except ValueError:
+            return urlparse("")
+    if url.startswith('://'):
+        url = f"{default_scheme}{url}"
+    elif url.startswith('//'):
+        url = f"{default_scheme}:{url}"
+    elif not url.startswith(('http://', 'https://', 'socks4://', 'socks4a://', 'socks5://', 'socks5h://')):
+        url = f"{default_scheme}://{url}"
+
+    parsed_url = urlparse(url)
+
+    if not parsed_url.scheme or not parsed_url.netloc:
+        return urlparse("")
+
+    return parsed_url
+
+
 class AdapterError(Exception):
     """Custom exception for adapter errors."""
 
     def __init__(self, code: int, msg: str) -> None:
-        """Initialize the AdapterError exception.
-
-        Args:
-            code (int): The error code.
-            msg (str): The error message.
-        """
         self.code = code
         self.msg = msg
 
     def __str__(self) -> str:
-        """Return the string representation of the exception.
-
-        Returns:
-            str: The error code and message.
-        """
         return f"{self.code}: {self.msg}"
 
 
 class WebsocketAdapter(BaseAdapter):
     """Adapter for handling WebSocket connections."""
 
-    def __init__(self, queue_size: int = 1024):
-        """Initialize the WebsocketAdapter.
+    ACTIVELY_CLOSE_REASON: bytes = b"KeyboardInterrupt"
 
-        Args:
-            queue_size (int): The size of the message queue.
-        """
+    def __init__(self, queue_size: int = 1024):
         super().__init__()
-        self._msg_queue = asyncio.Queue(queue_size)
+        self._msg_queue = queue.Queue(maxsize=queue_size)
         self._running = False
 
-        self._websocket: Optional[websockets.WebSocketClientProtocol] = None
-        # tasks
-        self._receive_task: Optional[asyncio.Task] = None
-        self._send_task: Optional[asyncio.Task] = None
-
-        # websocket loop
-        self._ws_loop = asyncio.new_event_loop()
-        self._ws_thread = threading.Thread(
-            target=self.start_loop, args=(self._ws_loop,), name="WSThread"
-        )
-        # websocket connection is running in an executor
-        self._ws_loop.set_default_executor(
-            ThreadPoolExecutor(max_workers=5, thread_name_prefix="WSAdapter")
-        )
-
-        # std
+        self._ws: Optional[websocket.WebSocket] = None
+        self._ws_thread = threading.Thread(target=self._run, name="WSThread")
         self._stdout = sys.stdout
         self._stdout_lock = threading.Lock()
 
-    @property
-    def running(self):
-        """Check if the websocket loop is running"""
-        return self._running
+        # ws info
+        self._close_code: Optional[int] = None
+        self._close_msg: Optional[str] = None
+
+        # store messages in bytes for download, httpie `--download` option
+        if sys.version_info >= (3, 12):
+            self._msgs_bytes = tempfile.NamedTemporaryFile(
+                mode="w+b", delete=True, delete_on_close=True
+            )
+        else:
+            self._msgs_bytes = tempfile.NamedTemporaryFile(mode="w+b", delete=True)
 
     @property
-    def connected(self):
-        """Check if the websocket is open"""
-        return self._websocket and self._websocket.open
+    def connected(self) -> bool:
+        return bool(self._ws and self._ws.connected)
 
-    def start_loop(self, thread_loop):
-        """
-        Start the loop in a separate thread
+    @property
+    def close_code(self) -> int:
+        return self._close_code or 0
+
+    @property
+    def close_msg(self) -> str:
+        return self._close_msg or ""
+
+    def convert2ws_headers(self, headers: Mapping) -> list[str]:
+        """Convert HTTP headers to WebSocket headers.
+
         Args:
-            thread_loop: asyncio event loop
-        """
-        asyncio.set_event_loop(thread_loop)
-        thread_loop.run_forever()
-
-    async def _connect(self, websocket_url: str) -> None:
-        """
-        Connect to the websocket if not already connected
-        Args:
-            websocket_url(str): the url of the websocket
-        Raises:
-            AdapterError: if the connection fails, raise AdapterError with code and msg
-        """
-        if not self._websocket or not self._websocket.open:
-            try:
-                self._websocket = await websockets.connect(websocket_url, close_timeout=4)
-            except OSError:
-                # Connect failed
-                raise AdapterError(500, "Cannot connect to websocket") from None
-            except websockets.InvalidStatusCode as e:
-                # Exceptions with status_code and custom msg
-                raise AdapterError(e.status_code, str(e)) from None
-            except websockets.AbortHandshake as e:
-                # Exceptions with status and body
-                raise AdapterError(e.status, e.body.decode()) from None
-            except (websockets.InvalidHandshake, websockets.InvalidURI) as e:
-                # Exceptions without status_code
-                raise AdapterError(500, str(e)) from None
-            except asyncio.TimeoutError:
-                # Connection timed out
-                raise AdapterError(500, "Connection timed out") from None
-
-    async def _receive(self):
-        """
-        Receive messages from the websocket
-        """
-        while self._running:
-            try:
-                message = await self._websocket.recv()
-                self._write_stdout(message)
-            except websockets.ConnectionClosed as e:
-                self._write_stdout(f"Connection closed when listening with code: {e.code}, reason: {e.reason}")
-                await self.put(None)
-                break
-            except RuntimeError as e:
-                # RuntimeError: If two coroutines call :meth:`recv` concurrently.
-                # Another coroutine is already receiving a message
-                # Should not happen since _receive is called only once.
-                self._write_stdout(str(e))
-                await self.put(None)
-                break
-            except asyncio.CancelledError:
-                break
-
-    async def _send_messages(self):
-        """
-        Send messages to the websocket
-        """
-        while self._running:
-            try:
-                message = await self._msg_queue.get()
-                if message is None:
-                    break
-                if message:
-                    if not isinstance(message, (bytes, str, Iterable, AsyncIterable)):
-                        # As a command-line program, this should not happen, just in case
-                        raise TypeError(
-                            f"Message must be bytes, str, Iterable, or AsyncIterable, got {type(message)}"
-                        )
-                    await self._websocket.send(message)
-            except websockets.ConnectionClosed as e:
-                self._write_stdout(f"Connection closed when sending with code: {e.code}, reason: {e.reason}")
-                break
-            except asyncio.CancelledError:
-                break
-
-    async def _run_tasks(self):
-        self._receive_task = asyncio.create_task(self._receive(), name="_receive_task")
-        self._send_task = asyncio.create_task(self._send_messages(), name="_send_task")
-        try:
-            await asyncio.gather(self._receive_task, self._send_task)
-        except Exception:  # noqa
-            # No stdout
-            pass
-
-    def read_stdin(self) -> Optional[str]:
-        """Read input from stdin synchronously.
+            headers (Mapping): HTTP headers.
 
         Returns:
-            Optional[str]: The input read from stdin, or None if no input is available.
+            list[str]: WebSocket headers.
         """
-        if not self._running:
-            return None
-        return _read_stdin()
+        if not headers:
+            return []
+        headers = dict(headers)
+        ws_headers = []
 
-    def send(self, request, **kwargs):
+        # Below headers already generated by websocket in handshake => _get_handshake_headers
+        # https://github.com/websocket-client/websocket-client/blob/master/websocket/_handshake.py#L83
+        ignore_keys = [
+            "upgrade",
+            "connection",
+            "origin",
+            "host",
+            "sec-websocket-key",
+            "sec-websocket-version",
+        ]
+        has_ua = False
+        for k, v in headers.items():
+            if k.lower() == "user-agent":
+                has_ua = True
+            if k.lower() in ignore_keys:
+                continue
+            if isinstance(v, bytes):
+                v = v.decode("utf-8")
+            ws_headers.append(f"{k}: {v}")
+        if not has_ua:
+            ws_headers.append(f"User-Agent: {headers.get('User-Agent', DEFAULT_UA)}")
+        return ws_headers
+
+    def _connect(self, request: PreparedRequest, **kwargs) -> None:
+        """Connect to the WebSocket if not already connected.
+
+        Args:
+            request (PreparedRequest): The request object.
+            **kwargs: Additional keyword arguments.
+        """
+        if self.connected:
+            return
+        options: dict[str, Any] = {}
+
+        proxy = kwargs.get("proxies")
+        if proxy:
+            ignored = []
+            proxy_url = None
+            for idx, (_, host_port) in enumerate(proxy.items()):
+                proxy_url = normalize_url(url=host_port, default_scheme="http")
+                if idx > 0:
+                    ignored.append(proxy_url.geturl())
+                    continue
+                options["http_proxy_host"] = proxy_url.hostname
+                options["http_proxy_port"] = proxy_url.port
+                options["proxy_type"] = proxy_url.scheme
+            if ignored:
+                msg = "\033[93mWarning: "
+                if proxy_url:
+                    msg += f"Using proxy {proxy_url.geturl()}. "
+                ignored = {i.decode("utf-8") if isinstance(i, bytes) else i for i in ignored}
+                msg += f"Proxy {', '.join(ignored)} is ignored because multiple proxies are not supported.\033[0m"
+                self._write_stdout(msg)
+
+        # --verify=yes/no
+        verify = kwargs.get("verify", True)
+        # IMPORTANT: httpie plugin not support specify ssl version and ciphers
+        cert: Optional[Union[HTTPieCertificate, str]] = kwargs.get("cert")
+        if not verify:
+            options["sslopt"] = {"cert_reqs": ssl.CERT_NONE}
+            if cert:
+                self._write_stdout(
+                    "\033[93mWarning: --cert is ignored because --verify is disabled.\033[0m"
+                )
+        else:
+            if isinstance(cert, HTTPieCertificate):
+                if cert.key_file:
+                    options["keyfile"] = Path(str(cert.key_file)).expanduser().resolve().as_posix()
+                if cert.key_file:
+                    options["certfile"] = (
+                        Path(str(cert.cert_file)).expanduser().resolve().as_posix()
+                    )
+                if cert.key_password:
+                    options["password"] = cert.key_password
+            elif isinstance(cert, str):
+                options["sslopt"] = {
+                    "cert_reqs": ssl.CERT_REQUIRED,
+                    "ca_certs": Path(cert).expanduser().resolve().as_posix(),
+                }
+        try:
+            self._ws: websocket.WebSocket = websocket.WebSocket(sslopt=options.get("sslopt"))
+            self._ws.connect(
+                request.url,
+                header=self.convert2ws_headers(request.headers),
+                redirect_limit=30,  # HTTPie default
+                http_proxy_host=options.get("http_proxy_host"),
+                http_proxy_port=options.get("http_proxy_port"),
+                proxy_type=options.get("proxy_type"),
+                keyfile=options.get("keyfile"),
+                certfile=options.get("certfile"),
+                password=options.get("password"),
+            )
+        except (websocket.WebSocketException, OSError) as e:
+            raise AdapterError(500, f"Cannot connect to websocket: {str(e)}") from None
+
+    def _receive(self):
+        """Receive messages from the WebSocket."""
+        while self._running and self.connected:
+            try:
+                resp_opcode, msg = self._ws.recv_data()
+                # print(f"""Received opcode: {resp_opcode}""", msg)
+                if resp_opcode == ABNF.OPCODE_CLOSE and len(msg) >= 2:
+                    # received a close message
+                    self._close_code = struct.unpack("!H", msg[0:2])[0]
+                    self._close_msg = msg[2:]
+                    if isinstance(self._close_msg, bytes):
+                        self._close_msg = self._close_msg.decode(encoding="utf8")
+                else:
+                    if isinstance(msg, bytes):
+                        msg = msg.decode("utf8")
+                    self._write_stdout(msg)
+            except websocket.WebSocketConnectionClosedException as e:
+                self._write_stdout(f"Connection closed: {str(e)}")
+                break
+            except OSError:
+                break
+
+    def _run(self):
+        """Run the WebSocket communication."""
+        try:
+            self._receive()
+        except Exception as e:
+            self._write_stdout(f"Error: {str(e)}")
+
+    def send(
+        self,
+        request,
+        stream=False,
+        timeout=None,
+        verify: Union[bool, str] = True,
+        cert=None,
+        proxies=None,
+    ):
         """Send a request to the WebSocket.
 
+        Will receive a string if only passed `--cert=path/to/cert`,
+        or a HTTPieCertificate object if `--cert-key` passed, like `--cert=path/to/cert --cert-key=path/to/key`
+
         Args:
-            request (PreparedRequest): The request to send.
+            request (PreparedRequest): The request object.
+            stream (bool): Whether to stream the response.
+            timeout (float): The timeout in seconds.
+            verify (Union[bool, str]): Whether to verify the SSL certificate.
+            cert (Union[HTTPieCertificate, str]): The path to the SSL certificate.
+            proxies (dict): The proxies to use.
 
         Returns:
-            Response: The response details from the WebSocket.
+            Response: Handshake response info.
         """
         self._running = True
-        self._ws_thread.start()
+        logger.debug(
+            f"ws connecting. stream: {stream}, verify: {verify}, proxy: {proxies}, timeout: {timeout}"
+        )
+        logger.debug(f"received headers: {request.headers}")
+
         try:
-            asyncio.run_coroutine_threadsafe(self._connect(request.url), self._ws_loop).result()
+            self._connect(
+                request, stream=stream, timeout=timeout, verify=verify, cert=cert, proxies=proxies
+            )
         except AdapterError as e:
             self.close()
             return self.dummy_response(request, e.code, e.msg)
 
+        self._ws_thread.start()
+        time.sleep(0.5)
         self._write_stdout(
             f"> Connected to {request.url}\n"
             "Type a message and press enter to send it\n"
@@ -226,71 +338,56 @@ class WebsocketAdapter(BaseAdapter):
         )
 
         try:
-            asyncio.run_coroutine_threadsafe(self._run_tasks(), self._ws_loop)
-
-            # Read input messages and put them in the queue
             while self._running and self.connected:
-                msg = self.read_stdin()
+                msg = _read_stdin()
                 if not msg:
                     continue
                 if not self._running or not self.connected:
-                    self._write_stdout(f"Websocket closed, message {msg} not sent")
+                    self._write_stdout(f"Websocket closed, message not sent: {msg}")
                     break
-                asyncio.run_coroutine_threadsafe(self.put(msg), self._ws_loop).result()
+                self.send_msg(msg)
         except KeyboardInterrupt:
-            self._write_stdout(
-                "\nKeyboardInterrupt received. Cleaning up, press Ctrl+C again to exit force."
-            )
-            # Ensure send message task quits
-            asyncio.run_coroutine_threadsafe(self.put(None), self._ws_loop).result()
+            self._write_stdout("\nOops! Disconnecting. Need to force quit? Press again!")
+            self._close_code = STATUS_ABNORMAL_CLOSED
+            self._close_msg = self.ACTIVELY_CLOSE_REASON.decode("utf8")
         finally:
             self.close()
+
         return self.dummy_response(request)
 
-    def _write_stdout(self, msg: AnyStr, newline: bool = True) -> None:
-        """
-        Write message to stdout
-        Args:
-            msg (AnyStr): The message to write.
-        """
+    def _write_stdout(self, msg: str, newline: bool = True) -> None:
+        """Write message to stdout."""
         if not self._running:
             return
+        if isinstance(msg, bytes):
+            msg = msg.decode("utf8")
         with self._stdout_lock:
             self._stdout.write(msg)
-            newline and self._stdout.write("\n")
+            if newline:
+                self._stdout.write("\n")
             self._stdout.flush()
 
     def dummy_response(
         self, request: PreparedRequest, status_code: int = 200, msg: str = ""
     ) -> Response:
-        """
-        Create a dummy response for requests send method
-        Args:
-            request (PreparedRequest): The request to respond to.
-            status_code (int): The status code of the response.
-            msg (str): The message to include in the response.
-
-        Returns:
-            Response: The dummy response.
-        """
+        """Create a dummy response for requests send method."""
         r = Response()
         r.status_code = status_code
         r.request = request
-        close_code = self._websocket.close_code if self._websocket else 1000
-        reason = self._websocket.close_reason if self._websocket else msg
-        headers = (
-            CaseInsensitiveDict(self._websocket.response_headers)
-            if self._websocket
-            else CaseInsensitiveDict({})
-        )
-        r.headers = headers
-        r._content = r.reason = reason
+        r.reason = msg
+        r.headers = CaseInsensitiveDict(self._ws.getheaders() if self._ws else {})
+        r._content = msg.encode("utf8") if msg else b""
         r.raw = io.BytesIO(
-            msg.encode("utf-8")
+            msg.encode("utf8")
             if msg
-            else f"""Close Code: {close_code}\nReason: {reason}""".encode("utf-8")
+            else (
+                f"Websocket connection info:\nClose Code: {self.close_code}\nClose Msg: {self.close_msg}".encode(
+                    "utf8"
+                )
+            )
         )
-        r.url = request.url
+        r.encoding = "utf-8"
+        r.url = request.url or ""
         return r
 
     def close(self) -> None:
@@ -298,35 +395,17 @@ class WebsocketAdapter(BaseAdapter):
         if self._running is False:
             return
         self._running = False
-        try:
-            asyncio.run_coroutine_threadsafe(self._close(), self._ws_loop).result()
-        except Exception:  # noqa
-            # No stacktrace in stdout
-            pass
-        finally:
-            self._ws_loop.call_soon_threadsafe(self._ws_loop.stop)
+        if self._ws and self._ws.connected:
+            self._ws.close(status=STATUS_ABNORMAL_CLOSED, reason=self.ACTIVELY_CLOSE_REASON)
+        if self._ws_thread.is_alive():
             self._ws_thread.join()
-            self._ws_loop.close()
 
-    async def _close(self):
-        if self._websocket and self._websocket.open:
-            await self._websocket.close(code=1000, reason="Closed by user")
-            await self._websocket.wait_closed()
-        for task in asyncio.all_tasks(self._ws_loop):
-            if not task.done():
-                task.cancel()
-                try:
-                    await task
-                except asyncio.CancelledError:
-                    pass
-
-    async def put(self, msg: Optional[AnyStr]) -> None:
-        """Put a message into the message queue.
-
-        Args:
-            msg (Optional[AnyStr]): The message to put into the queue.
-        """
-        await self._msg_queue.put(msg)
+    def send_msg(self, message: str) -> int:
+        if not self._ws:
+            raise requests.exceptions.RequestException("WebSocket not initialized")
+        length: int = self._ws.send_text(message)
+        logger.debug(f"Sent message: {message}, frame length: {length}")
+        return length
 
 
 class BaseWebsocketPlugin(TransportPlugin):
@@ -336,11 +415,6 @@ class BaseWebsocketPlugin(TransportPlugin):
     description = "WebSocket transport plugin for HTTPie"
 
     def get_adapter(self) -> WebsocketAdapter:
-        """Get the WebSocket adapter.
-
-        Returns:
-            WebsocketAdapter: The WebSocket adapter.
-        """
         return WebsocketAdapter()
 
 
@@ -360,15 +434,27 @@ class WebsocketSPlugin(BaseWebsocketPlugin):
 
 if __name__ == "__main__":
     import argparse
+
     import requests
 
     parser = argparse.ArgumentParser(prog="python -m httpie_websocket")
     parser.add_argument("url")
+    parser.add_argument("--proxy", help="proxy url")
     args = parser.parse_args()
+    print(args.proxy)
+
+    proxy_u = urlparse(args.proxy)
+    proxy = {"http": proxy_u.geturl(), "https": proxy_u.geturl()}
 
     adapter = WebsocketAdapter()
     session = requests.Session()
     session.mount("ws://", adapter)
     session.mount("wss://", adapter)
-    session.request("WEBSOCKET", args.url)
+    # resp = session.request("WEBSOCKET", args.url, proxies={"socks5": "//195.209.188.101:58543"})
+    resp = session.request("WEBSOCKET", args.url, proxies=proxy)
+    print("\n".join(f"{k}: {v}" for k, v in resp.headers.items()), end="\n\n")
+    try:
+        print(json.dumps(resp.text, indent=4))
+    except requests.JSONDecodeError:
+        print(resp.text)
     session.close()